+++ cases-extra/testcases/faulty_defmacro_consumer.clj
@@ -1,0 +1,5 @@
+(ns testcases.faulty-defmacro-consumer
+  (:require
+   [testcases.faulty-defmacro]))
+
+(def x (testcases.faulty-defmacro/dubious 2))