+++ cases/testcases/faulty_defmacro.clj
@@ -1,0 +1,6 @@
+(ns testcases.faulty-defmacro)
+
+(defmacro dubious
+  "A macro that will cause a linter fault"
+  [x]
+  `(-> ~x))